+++ packages/lumx-react/src/components/popover-dialog/PopoverDialog.stories.tsx
@@ -1,13 +1,13 @@
-import { mdiMenu } from '@lumx/icons/index';
-import { mdiSettings } from '@lumx/icons/v4-to-v5-aliases';
 import React, { useRef, useState } from 'react';
+import { mdiMenu, mdiSettings } from '@lumx/icons';
+import { Placement } from '@lumx/react';
+
 import { PopoverDialog, PopoverDialogProps } from '.';
 import { Emphasis, Orientation, Size, Typography } from '..';
 import { Button, IconButton } from '../button';
 import { FlexBox } from '../flex-box';
 import { Heading } from '../heading';
 import { List, ListItem } from '../list';
-import { Placement } from '../popover/Popover';
 import { Toolbar } from '../toolbar';
 
 const WithButton = (Story: any, context: any) => {