+++ packages/lumx-react/src/components/popover/index.ts
@@ -1,1 +1,3 @@
 export * from './Popover';
+export type { Offset, Elevation } from './constants';
+export { Placement } from './constants';